[Frontend] Refactored LocationPage
