+++ frontend/src/components/LocationMap.js
@@ -1,44 +1,26 @@
-import {GoogleMap, useLoadScript} from "@react-google-maps/api";
-import React from "react";
+import {IconButton} from "@material-ui/core";
+import ListIcon from "@material-ui/icons/List";
+import GoogleMapsContainer from "./GoogleMapsContainer";
 import styled from "styled-components/macro";
-import LocationMarker from "./LocationMarker";
 
-export default function LocationMap({locations}) {
-
-    const googleLibraries = []
-    const mapCenter = {
-        lat: 51.163361,
-        lng: 10.447683,
-    }
-    const mapContainerStyle = {
-        width: '80vw',
-        height: '80vh',
-    }
-
-    const {isLoaded, loadError} = useLoadScript({
-        googleMapsApiKey: process.env.GOOGLE_MAPS_API_KEY,
-        googleLibraries,
-    })
-
-    if (loadError) return "Error";
-    if (!isLoaded) return "Loading...";
+export default function LocationMap({locations, switchView}) {
 
     return (
-        <Wrapper>
-            <GoogleMap
-                mapContainerStyle={mapContainerStyle}
-                zoom={5}
-                center={mapCenter}>
-                {locations.map((location) =>
-                    <LocationMarker key={location.id} position={{lat: location.lat, lng: location.lng}}/>
-                    )
-                }
-            </GoogleMap>
-        </Wrapper>
+        <>
+            <Wrapper>
+                <IconButton onClick={(event) => {
+                    switchView(event, false)
+                }} aria-label="list" variant="outlined">
+                    <ListIcon/>
+                </IconButton>
+            </Wrapper>
+            <GoogleMapsContainer locations={locations}/>
+        </>
     )
+
 }
 
 const Wrapper = styled.div`
-  display: flex;
-  justify-content: center;
+display: flex;
+justify-content: end;
 `
